refactor: BlackJackCondition 상수 위치 수정
