+++ src/main/java/domain/PlayingCardValue.java
@@ -1,6 +1,6 @@
 package domain;
 
-import static domain.constant.GameOption.BLACKJACK_CONDITION;
+import static domain.BlackJackGame.BLACKJACK_CONDITION;
 
 public enum PlayingCardValue {
     TWO(2), THREE(3), FOUR(4), FIVE(5), SIX(6), SEVEN(7), EIGHT(8), NINE(9), TEN(10), JACK(10), QUEEN(10), KING(10), ACE(11);
