--- src/main/java/domain/constant/GameOption.java
@@ -1,5 +1,0 @@
-package domain.constant;
-
-public class GameOption {
-    public static final int BLACKJACK_CONDITION = 21;
-}
